Fixed hours tens digit non-clearing bug
When the clock would turn from 12:59 to 1:00, the first 1 would not be
erased, and it would say 11:00 instead of 1:00.     It is now fixed so
that the first 1 would be erased.
I thought of only doing it at the time change of 12:59 to 1:00, but then
what if  it is 12:50 and the computer is woke up at 3:50?  The 1 would
still be there.